Update react-scripts
This is to update a nested dependency, node-notifier, which has known
vulnerabilities. Note the node-notifier vulnerabilites do not affect the
application once deployed, so this doesn't warrant a new release.

The changelogs for react-scripts indicate no breaking changes between
these versions, my testing has been to test building the application
which succeeds.